Added support for checking learnable collection bundles.
These kind of items are considered "learnable" and collectable, like battle pets, toys, etc. Collections contain a set of multiple items, and if all are collected, the collection is considered complete and the item grayed out as "Already known" to make it easier to filter and see in the list.

The "Learnable" and "Learnable: Collected" filters will consider these kind of bundles.