Track the event for merchant show/closed by setting the `merchantOpen` boolean accordingly.
Reply on this new boolean in `UpdateMerchantInfo` so that if we're not on a merchant we ensure to cleanup and answer properly, we can't be on a merchant if the event said we closed the interface.

Added BuyEmAll support, if the addon is loaded, it will be used for stack purchase instead of the built-in standard frame.

The new Open/Close/IsOpen methods on the quantity button widget will use the API for BuyEmAll when available, otherwise fallback to default behavior.