Merchant filter needs to track the updatesd on a per-item basis and not just as "the last one gets updated" like it previously was.
This wasn't really noticeable as long a full update was performed, which often happens once filter events settle, but still it could lead to some oddities.

Added a loading spinner when the frame shows until it is done loading data. Would only affect the frame when initially shown as the server needs a bit of time to send data to the client.

Added status text when the vendor has no items to sell, or is filtered by the filter system, so the user knows it's empty and not that it's waiting to load data (though the spinner now helps cover the loading aspect).