Fixed tooltip scanning issues around profession scanning of requirements.
An item can have multiple requirements, and we mostly care about the profession one for coloring, and the dropdown filters care for all of them combined.

There is more work to be done to confirm the other kind of requirements: level, rating, achievement, guild, reputation, specialization, renown, and check for other kinds that we should add to the filter and scanning code.