Renamed the old `CanLean` method to a universal `HasRequirements` method which will scan and extract the various requirements, the old method did the same logic just worse and used an older naming convention which is now corrected.
Requirements are scanned from the bottom up, and will stop scanning as soon a non-requirement is encountered after any requirements that were found.

This is to avoid embed issues, where the tooltip contained data from an embeded item, but we couldn't know where it ends and the regular tooltip continues.

Updated the stats filter and the requirements filter, it had to be adjusted to consider the checked options/values, so only those were taken into account when deciding if to show or hide the item in question.

There are still some oddities with the requirements, will need to do more testing.