The merchant item has awareness on which button is using it.
Moved the cleanup of button code in a view factory resetter handler.

The tooltip scan will ensure the results still match the expected merchant item ID, and abort if not. (This could happen if a response is delayed and the frame reused for another item in the meantime.)

When the view assigns a button it also assigns that buttons merchant item object, and the parentButton on said object. (And during resetting this is cleaned up.)

The update merchant routine will, when possible, call the `UpdateMerchantItemButton` method as data is refreshed so the visuals match the backend state.

Previously this only occured when the button is displayed, or when specific events related to bags occured. This would sometimes lead to state missmatch, but hopefully not anymore.