Split up the `tooltipData` into additional props to track state and data and the unique scanner ID (for response tracking).
Moved the response handling and callbacks to their own methods `RefreshTooltipData` and `OnTooltipScanResponse` which are called when appropriate.

This simplified the code in the `Refresh` routine to simply use the `tooltipData` when available, otherwise start a scan task and assign the relevant fields and provide the callbacks needed to update the correct item once the response returns.

The `TooltipScanner` is now specifically containing the relevant methods we need for its purpose.