The merchant index also has a `hasIndex` boolean for easy checking.
Most places where the index was used, also now check if it's valid. This only applies to situations where the merchant is still loading data.

The rows that display this rare loading state won't appear like items, it will hide all other behavior, and simply show "Loading..." until the data is available for it to display itself.