The stat filter can miss some stat names due the global space not containing the localization as expected.
Created a simple mapping table to redirect the key findings to an existing localized stat name.

This needs to be checked on classic era and mist clients, but for mainline it works well.

The `ReMapStatTable` func was implemented to override duplicate entries that appeared for some reason, into one filter value to avoid duplication.

This was specific to the resistances, so I used the `statTextMap` to find the keys and map them to the values like with the translations.

This fixed the filter and deduped the dropdown options.