Implemented merchant tooltip scanning, but the issue is additional async delays with the special requirement text. Not to mention the localization issue, most those texts have no globalstrings to use to pattern match.
The hyperlink scan approach is still the active one, though the code supports to swap the logic to the merchant style when we get it to a usable state.

Small improvements to the requirements filter and the renown presentation (from my tests, but there are probably many other kind of requirements).